refactor: move test helpers and fixtures under src/testing/
