+++ src/task-utils.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, mock, test } from "bun:test";
 import type { CoderClient } from "./coder-client";
 import { CoderAPIError } from "./coder-client";
 import { TestLogger } from "./infra/logger";
-import { MockCoderClient, mockTask } from "./test-helpers";
+import { MockCoderClient, mockTask } from "./testing/helpers";
 import {
 	generateTaskName,
 	lookupAndEnsureActiveTask,
