+++ src/testing/helpers.ts
@@ -1,7 +1,7 @@
 import { mock } from "bun:test";
-import type { CoderClient, ExperimentalCoderSDKTask } from "./coder-client";
-import { TaskIdSchema, TaskNameSchema } from "./coder-client";
-import type { GitHubClient } from "./github-client";
+import type { CoderClient, ExperimentalCoderSDKTask } from "../coder-client";
+import { TaskIdSchema, TaskNameSchema } from "../coder-client";
+import type { GitHubClient } from "../github-client";
 
 // ── Mock Task Data ──────────────────────────────────────────────────────────
